commit ea6920ffcac2b0053ce0694b6bcf8562ae006f6f
    15948 node enumeration should have property groups
    15944 want fac_prov_libhp
commit a941314348920b2bdf3790bc98861675ee18d6ec
    15946 want PCIe LED mechanism for systems that aren't strictly ExpressModule
    15945 pciehpc slot properties need a way to go back to the default
    15947 pciehpc should be able to set power led
commit bd97c7ce2344fa3252d8785c35895490916bc79b
    15587 PCI bridge subsystem IDs are ignored
    15707 pci-hotplug-type is redundant
    15708 class_pci_items should be in a header, unsigned
    15709 x86 pcicfg property logic has several problems
commit 350b21149ccca29970f4e8cd25afd3aa6b238887
    15942 PCIe hotplug insertion does not configure the slot
    15943 PCIe hotplug startup races, leaving configuration incomplete
    15941 cfgadm -c configure does nothing for enabled devices
commit 9c84c2e828ea91e81ac1287dbd9e906252c84df0
    15940 pcie module is platform-specific
commit e3c4d829fd2f5ea07a7203c4a9a02f1b8b3b18d6
    11466 remove old closed iconv tables
commit 6990962ce8f191dd6bb6a174a2f3dec3e3a51f18
    15964 ZFS test fixes to go with 15338
commit 705610572a32b33c0da7d60b05c6dc30597a9e5a
    15338 Implicit owner rights should be configurable

Conflicts:
    manifest
    usr/src/man/man8/zfs.8

+++ usr/src/common/pci/pci_strings.c
@@ -235,4 +235,4 @@ const pci_class_strings_t class_pci[] = {
 	{ 19, 0, 0,	"Non-essential Instrumentation",	"unknown" },
 };
 
-int class_pci_items = sizeof (class_pci) / sizeof (class_pci[0]);
+size_t class_pci_items = sizeof (class_pci) / sizeof (class_pci[0]);
